+++ SetupConfig.py
@@ -55,6 +55,7 @@ classifiers = [line.strip() for line in '''\
   Programming Language :: Python :: 3.5
   Programming Language :: Python :: 3.6
   Programming Language :: Python :: 3.7
+  Programming Language :: Python :: 3.8
   Topic :: Internet :: WWW/HTTP
   Topic :: Internet :: WWW/HTTP :: Dynamic Content
   Topic :: Internet :: WWW/HTTP :: Site Management